Added templated setData function to ShaderCursor to make it easier to set simple uniform values; Added a new compareComputeResult function to gfx-test-util for comparing texture resource; Expanded getCUDAFormat in render-cuda to support more formats; Reorganized the shared texture test to support adding tests for other formats later, test to check that the shared texture can be used in a shader now works

+++ tools/gfx-unit-test/gfx-test-util.h
@@ -30,7 +30,7 @@ namespace gfx_test
         gfx::IDevice* device,
         gfx::ITextureResource* texture,
         gfx::ResourceState state,
-        uint8_t* expectedResult,
+        float* expectedResult,
         size_t expectedBufferSize);
 
         /// Reads back the content of `buffer` and compares it against `expectedResult` with a set tolerance.
